Combine classify and collect python client Fixes #14
