+++ uncat/main.py
@@ -1,0 +1,88 @@
+#Dependancies
+import cv2
+import numpy as np
+import os
+from matplotlib import pyplot as plt
+import time
+import mediapipe as mp
+from tensorflow.keras.models import Sequential
+
+import draw
+import SVmodel
+
+#Keypoints Using MP Holistic
+mp_holistic = mp.solutions.holistic # Holistic model
+
+# Create and initialize model
+model = SVmodel.SVModel()
+
+def mediapipe_detection(image, model):
+    image = cv2.cvtColor(image, cv2.COLOR_BGR2RGB) # COLOR CONVERSION BGR 2 RGB
+    image.flags.writeable = False                  # Image is no longer writeable
+    results = model.process(image)                 # Make prediction
+    image.flags.writeable = True                   # Image is now writeable 
+    image = cv2.cvtColor(image, cv2.COLOR_RGB2BGR) # COLOR COVERSION RGB 2 BGR
+    return image, results
+
+def extract_keypoints(results):
+    pose = np.array([[res.x, res.y, res.z, res.visibility] for res in results.pose_landmarks.landmark]).flatten() if results.pose_landmarks else np.zeros(33*4)
+    face = np.array([[res.x, res.y, res.z] for res in results.face_landmarks.landmark]).flatten() if results.face_landmarks else np.zeros(468*3)
+    lh = np.array([[res.x, res.y, res.z] for res in results.left_hand_landmarks.landmark]).flatten() if results.left_hand_landmarks else np.zeros(21*3)
+    rh = np.array([[res.x, res.y, res.z] for res in results.right_hand_landmarks.landmark]).flatten() if results.right_hand_landmarks else np.zeros(21*3)
+    return np.concatenate([pose, face, lh, rh])
+
+#Main
+sequence = []
+sentence = []
+threshold = 0.3
+cap = cv2.VideoCapture(1)
+# Set mediapipe model 
+with mp_holistic.Holistic(min_detection_confidence=0.5, min_tracking_confidence=0.5) as holistic:
+    while cap.isOpened():
+
+        # Read feed
+        ret, frame = cap.read()
+
+        # Make detections
+        image, results = mediapipe_detection(frame, holistic)
+        
+        # Draw landmarks
+        draw.styled_landmarks(image, results)
+        
+        # 2. Prediction logic
+        keypoints = extract_keypoints(results)
+#         sequence.insert(0,keypoints)
+#         sequence = sequence[:30]
+        sequence.append(keypoints)
+        sequence = sequence[-30:]
+        
+        if len(sequence) == 30:
+            res = model.predict(sequence)
+            print(model.actions[np.argmax(res)])
+            print(res)
+            
+            
+        #3. Viz logic
+            if res[np.argmax(res)] > threshold: 
+                if len(sentence) > 0: 
+                    if model.actions[np.argmax(res)] != sentence[-1]:
+                        sentence.append(model.actions[np.argmax(res)])
+                else:
+                    sentence.append(model.actions[np.argmax(res)])
+
+            if len(sentence) > 5: 
+                sentence = sentence[-5:]
+            
+        cv2.rectangle(image, (0,0), (640, 40), (245, 117, 16), -1)
+        cv2.putText(image, ' '.join(sentence), (3,30), 
+                       cv2.FONT_HERSHEY_SIMPLEX, 1, (255, 255, 255), 2, cv2.LINE_AA)
+        
+        # Show to screen
+        cv2.imshow('OpenCV Feed', image)
+
+        # Break gracefully
+        if cv2.waitKey(10) & 0xFF == ord('q'):
+            break
+    cap.release()
+    cv2.destroyAllWindows()
+
